+++ lib/protoboeuf.rb
@@ -3,4 +3,5 @@
 module ProtoBoeuf
   autoload :CodeGen, "protoboeuf/codegen"
   autoload :Google, "protoboeuf/google"
+  autoload :ProtocUtils, "protoboeuf/protoc_utils"
 end
