+++ test/codegen/field_test.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require "helper"
-require "protoboeuf/codegen/field"
 
 module ProtoBoeuf
   class CodeGen
